Handle deprecations. Remove a few deprecated classes (they have replacements).
