+++ src/smartcontracts/transactionPayloadBuilders.spec.ts
@@ -1,16 +1,15 @@
 import { assert } from "chai";
 import { Address } from "../address";
-import { TransactionPayload } from "../transactionPayload";
 import { ContractFunction } from "./function";
 import { Code } from "./code";
 import { CodeMetadata } from "./codeMetadata";
 import { AddressValue, U32Value } from "./typesystem";
+import { ContractCallPayloadBuilder, ContractDeployPayloadBuilder, ContractUpgradePayloadBuilder } from "./transactionPayloadBuilders";
 
 
 describe("test contract payload builders", () => {
     it("should prepare deploy correctly", async () => {
-        let payload = TransactionPayload
-            .contractDeploy()
+        let payload = new ContractDeployPayloadBuilder()
             .setCode(Code.fromBuffer(Buffer.from([1, 2, 3, 4])))
             .setCodeMetadata(new CodeMetadata(true, false, true))
             .addInitArg(new U32Value(1024))
@@ -20,8 +19,7 @@ describe("test contract payload builders", () => {
     });
 
     it("should prepare upgrade correctly", async () => {
-        let payload = TransactionPayload
-            .contractUpgrade()
+        let payload = new ContractUpgradePayloadBuilder()
             .setCode(Code.fromBuffer(Buffer.from([1, 2, 3, 4])))
             .setCodeMetadata(new CodeMetadata(true, false, true))
             .addInitArg(new U32Value(1024))
@@ -32,8 +30,7 @@ describe("test contract payload builders", () => {
 
     it("should prepare call correctly", async () => {
         let alice = new Address("erd1l453hd0gt5gzdp7czpuall8ggt2dcv5zwmfdf3sd3lguxseux2fsmsgldz");
-        let payload = TransactionPayload
-            .contractCall()
+        let payload = new ContractCallPayloadBuilder()
             .setFunction(new ContractFunction("transferToken"))
             .addArg(new AddressValue(alice))
             .addArg(new U32Value(1024))
